app/requirements.txt
flask==2.0.1
boto3==1.26.137
python-dotenv==0.19.0
gunicorn==20.1.0 